+++ src/models/trakt/trakt-authentication.model.ts
