+++ src/models/trakt/trakt-calendar.model.ts
@@ -1,6 +1,6 @@
-import type { TraktEpisode } from '~/models/trakt-episode.model';
-import type { TraktMovie } from '~/models/trakt-movie.model';
-import type { TraktShow } from '~/models/trakt-show.model';
+import type { TraktEpisode } from '~/models/trakt/trakt-episode.model';
+import type { TraktMovie } from '~/models/trakt/trakt-movie.model';
+import type { TraktShow } from '~/models/trakt/trakt-show.model';
 
 export type TraktCalendarShow = {
   /** Timestamp in ISO 8601 GMT format (YYYY-MM-DD'T'hh:mm:ss.sssZ) */
