+++ src/models/trakt/trakt-like.model.ts
@@ -1,4 +1,4 @@
-import type { TraktUser } from '~/models/trakt-user.model';
+import type { TraktUser } from '~/models/trakt/trakt-user.model';
 
 export type TraktLike = {
   /** Timestamp in ISO 8601 GMT format (YYYY-MM-DD'T'hh:mm:ss.sssZ) */
