+++ src/models/trakt/trakt-season.model.ts
@@ -1,6 +1,6 @@
-import type { TraktEpisode } from '~/models/trakt-episode.model';
-import type { TraktApiIds } from '~/models/trakt-id.model';
-import type { TraktCast } from '~/models/trakt-people.model';
+import type { TraktEpisode } from '~/models/trakt/trakt-episode.model';
+import type { TraktApiIds } from '~/models/trakt/trakt-id.model';
+import type { TraktCast } from '~/models/trakt/trakt-people.model';
 
 export type TraktSeasonShort = {
   number: number;
