+++ src/models/trakt/trakt-user.model.ts
@@ -1,9 +1,9 @@
-import type { TraktComment, TraktCommentMedia } from '~/models/trakt-comment.model';
-import type { TraktApiIds } from '~/models/trakt-id.model';
-import type { TraktList } from '~/models/trakt-list.model';
-import type { TraktMovie } from '~/models/trakt-movie.model';
-import type { TraktSeason } from '~/models/trakt-season.model';
-import type { TraktShow } from '~/models/trakt-show.model';
+import type { TraktComment, TraktCommentMedia } from '~/models/trakt/trakt-comment.model';
+import type { TraktApiIds } from '~/models/trakt/trakt-id.model';
+import type { TraktList } from '~/models/trakt/trakt-list.model';
+import type { TraktMovie } from '~/models/trakt/trakt-movie.model';
+import type { TraktSeason } from '~/models/trakt/trakt-season.model';
+import type { TraktShow } from '~/models/trakt/trakt-show.model';
 import type { RequireAtLeastOne } from '~/utils/typescript.utils';
 
 export type TraktPrivateUser = {
