+++ src/models/trakt/trakt-watchlist.model.ts
@@ -1,8 +1,8 @@
-import type { TraktEpisode } from '~/models/trakt-episode.model';
-import type { TraktList } from '~/models/trakt-list.model';
-import type { TraktMovie } from '~/models/trakt-movie.model';
-import type { TraktSeason } from '~/models/trakt-season.model';
-import type { TraktShow } from '~/models/trakt-show.model';
+import type { TraktEpisode } from '~/models/trakt/trakt-episode.model';
+import type { TraktList } from '~/models/trakt/trakt-list.model';
+import type { TraktMovie } from '~/models/trakt/trakt-movie.model';
+import type { TraktSeason } from '~/models/trakt/trakt-season.model';
+import type { TraktShow } from '~/models/trakt/trakt-show.model';
 
 type BaseTraktWatchlistItem<E extends 'extended' | 'short' | 'any' = 'any'> = {
   movie: TraktMovie<E>;
