+++ src/services/trakt-client/api/endpoints/episodes.endpoint.ts
@@ -1,12 +1,12 @@
-import type { TraktComment } from '~/models/trakt-comment.model';
-import type { TraktRating, TraktTranslation } from '~/models/trakt-entity.model';
-import type { TraktEpisode, TraktEpisodeCast, TraktShowStats } from '~/models/trakt-episode.model';
+import type { TraktComment } from '~/models/trakt/trakt-comment.model';
+import type { TraktRating, TraktTranslation } from '~/models/trakt/trakt-entity.model';
+import type { TraktEpisode, TraktEpisodeCast, TraktShowStats } from '~/models/trakt/trakt-episode.model';
 
-import type { TraktList } from '~/models/trakt-list.model';
+import type { TraktList } from '~/models/trakt/trakt-list.model';
 
-import type { TraktUser } from '~/models/trakt-user.model';
+import type { TraktUser } from '~/models/trakt/trakt-user.model';
 
-import { TraktApiExtended, type TraktApiParamsExtended, type TraktApiParamsPagination, TraktClientEndpoint } from '~/models/trakt-client.model';
+import { TraktApiExtended, type TraktApiParamsExtended, type TraktApiParamsPagination, TraktClientEndpoint } from '~/models/trakt/trakt-client.model';
 import { HttpMethod } from '~/utils/http.utils';
 
 type BaseEpisodeParam = {
