+++ src/services/trakt-client/api/endpoints/lists.endpoint.ts
@@ -1,10 +1,10 @@
-import type { TraktApiParamsExtended, TraktApiParamsPagination } from '~/models/trakt-client.model';
+import type { TraktApiParamsExtended, TraktApiParamsPagination } from '~/models/trakt/trakt-client.model';
 
-import type { TraktComment } from '~/models/trakt-comment.model';
-import type { TraktLike } from '~/models/trakt-like.model';
-import type { TraktList, TraktListItem, TraktListList } from '~/models/trakt-list.model';
+import type { TraktComment } from '~/models/trakt/trakt-comment.model';
+import type { TraktLike } from '~/models/trakt/trakt-like.model';
+import type { TraktList, TraktListItem, TraktListList } from '~/models/trakt/trakt-list.model';
 
-import { TraktApiExtended, TraktClientEndpoint } from '~/models/trakt-client.model';
+import { TraktApiExtended, TraktClientEndpoint } from '~/models/trakt/trakt-client.model';
 
 import { HttpMethod } from '~/utils/http.utils';
