+++ src/services/trakt-client/api/endpoints/people.endpoint.ts
@@ -1,11 +1,11 @@
-import type { TraktList } from '~/models/trakt-list.model';
-import type { TraktMovieCast } from '~/models/trakt-movie.model';
-import type { TraktPerson, TraktPersonUpdate } from '~/models/trakt-people.model';
+import type { TraktList } from '~/models/trakt/trakt-list.model';
+import type { TraktMovieCast } from '~/models/trakt/trakt-movie.model';
+import type { TraktPerson, TraktPersonUpdate } from '~/models/trakt/trakt-people.model';
 
-import type { TraktShowCast } from '~/models/trakt-show.model';
+import type { TraktShowCast } from '~/models/trakt/trakt-show.model';
 
-import { TraktApiExtended, type TraktApiParamsExtended, type TraktApiParamsPagination, TraktClientEndpoint } from '~/models/trakt-client.model';
-import { type StartDateParam, transformStartDate, validateStartDate } from '~/models/trakt-entity.model';
+import { TraktApiExtended, type TraktApiParamsExtended, type TraktApiParamsPagination, TraktClientEndpoint } from '~/models/trakt/trakt-client.model';
+import { type StartDateParam, transformStartDate, validateStartDate } from '~/models/trakt/trakt-entity.model';
 import { HttpMethod } from '~/utils/http.utils';
 
 export const people = {
