+++ src/services/trakt-client/api/endpoints/shows.endpoint.ts
@@ -1,7 +1,7 @@
-import type { TraktComment } from '~/models/trakt-comment.model';
-import type { TraktEpisode } from '~/models/trakt-episode.model';
-import type { TraktList } from '~/models/trakt-list.model';
-import type { TraktCollectionProgress, TraktProgressReset, TraktWatchedProgress } from '~/models/trakt-progress.model';
+import type { TraktComment } from '~/models/trakt/trakt-comment.model';
+import type { TraktEpisode } from '~/models/trakt/trakt-episode.model';
+import type { TraktList } from '~/models/trakt/trakt-list.model';
+import type { TraktCollectionProgress, TraktProgressReset, TraktWatchedProgress } from '~/models/trakt/trakt-progress.model';
 import type {
   TraktShow,
   TraktShowAnticipated,
@@ -12,9 +12,9 @@ import type {
   TraktShowStats,
   TraktShowTrending,
   TraktShowUpdate,
-} from '~/models/trakt-show.model';
+} from '~/models/trakt/trakt-show.model';
 
-import type { TraktUser } from '~/models/trakt-user.model';
+import type { TraktUser } from '~/models/trakt/trakt-user.model';
 
 import {
   TraktApiExtended,
@@ -23,7 +23,7 @@ import {
   type TraktApiParamsPagination,
   type TraktApiTemplateOptions,
   TraktClientEndpoint,
-} from '~/models/trakt-client.model';
+} from '~/models/trakt/trakt-client.model';
 import {
   type StartDateParam,
   type TraktAlias,
@@ -32,7 +32,7 @@ import {
   type TraktTranslation,
   transformStartDate,
   validateStartDate,
-} from '~/models/trakt-entity.model';
+} from '~/models/trakt/trakt-entity.model';
 import { type TraktApiShowFilters, TraktApiShowFilterValues } from '~/services/trakt-client/api/trakt-api.filters';
 import { HttpMethod } from '~/utils/http.utils';
